fix v11- Types from momgoose

+++ src/repositories/users-db-repository.ts
@@ -1,5 +1,5 @@
 import { ObjectId } from 'mongodb'
-
+import { Types } from 'mongoose'
 import {
   type UserDBType,
   type ViewUserType,
@@ -248,16 +248,17 @@ export const usersRepository = {
     return result
   },
   async delete(id: string): Promise<boolean | null> {
-    // try {
-    const result = await UserModel.deleteOne({ _id: new ObjectId(id) })
-    console.log('258++users.repo--результат удвления юзера', result)
+    try {
+      if (!Types.ObjectId.isValid(id)) return null
+      const result = await UserModel.deleteOne({ _id: new ObjectId(id) })
+      console.log('258++users.repo--результат удвления юзера', result)
 
-    return result.deletedCount === 1
-    // } catch (e) {
-    // console.log('257---catch', e)
+      return result.deletedCount === 1
+    } catch (e) {
+      console.log('257---catch', e)
 
-    // return null
-    // }
+      return null
+    }
   },
 
   async create(newElement: UserDBType): Promise<string> {
